+++ CHANGELOG.md
@@ -6,8 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
-- Performance improve on change/lazy items rendering.
+### Added
+- Lazy rendering to items.
+- lazyRenderHeight and lazyRenderOffset props.
 
 ## [0.27.2] - 2020-12-21